Fixed type of abs() function in connection.c

+++ src/connection.c
@@ -383,7 +383,7 @@ int mapper_boundary_perform(mapper_connection connection,
                     break;
                 case BA_FOLD:
                     // fold value around range minimum
-                    difference = fabsf(value - dest_min);
+                    difference = fabs(value - dest_min);
                     value = dest_min + difference;
                     if (value > dest_max) {
                         // value now exceeds range maximum!
@@ -398,7 +398,7 @@ int mapper_boundary_perform(mapper_connection connection,
                                 break;
                             case BA_FOLD:
                                 // both boundary actions are set to fold!
-                                difference = fabsf(value - dest_max);
+                                difference = fabs(value - dest_max);
                                 modulo_difference = difference
                                     - ((int)(difference / total_range)
                                        * total_range);
@@ -410,7 +410,7 @@ int mapper_boundary_perform(mapper_connection connection,
                                 break;
                             case BA_WRAP:
                                 // wrap value back from range minimum
-                                difference = fabsf(value - dest_max);
+                                difference = fabs(value - dest_max);
                                 modulo_difference = difference
                                     - ((int)(difference / total_range)
                                        * total_range);
@@ -423,7 +423,7 @@ int mapper_boundary_perform(mapper_connection connection,
                     break;
                 case BA_WRAP:
                     // wrap value back from range maximum
-                    difference = fabsf(value - dest_min);
+                    difference = fabs(value - dest_min);
                     modulo_difference = difference
                         - (int)(difference / total_range) * total_range;
                     value = dest_max - modulo_difference;
@@ -445,7 +445,7 @@ int mapper_boundary_perform(mapper_connection connection,
                     break;
                 case BA_FOLD:
                     // fold value around range maximum
-                    difference = fabsf(value - dest_max);
+                    difference = fabs(value - dest_max);
                     value = dest_max - difference;
                     if (value < dest_min) {
                         // value now exceeds range minimum!
@@ -460,7 +460,7 @@ int mapper_boundary_perform(mapper_connection connection,
                                 break;
                             case BA_FOLD:
                                 // both boundary actions are set to fold!
-                                difference = fabsf(value - dest_min);
+                                difference = fabs(value - dest_min);
                                 modulo_difference = difference
                                     - ((int)(difference / total_range)
                                        * total_range);
@@ -472,7 +472,7 @@ int mapper_boundary_perform(mapper_connection connection,
                                 break;
                             case BA_WRAP:
                                 // wrap value back from range maximum
-                                difference = fabsf(value - dest_min);
+                                difference = fabs(value - dest_min);
                                 modulo_difference = difference
                                     - ((int)(difference / total_range)
                                        * total_range);
@@ -485,7 +485,7 @@ int mapper_boundary_perform(mapper_connection connection,
                     break;
                 case BA_WRAP:
                     // wrap value back from range minimum
-                    difference = fabsf(value - dest_max);
+                    difference = fabs(value - dest_max);
                     modulo_difference = difference
                         - (int)(difference / total_range) * total_range;
                     value = dest_min + modulo_difference;
